feat: Add start button and flashing functionality
This commit introduces a 'Start' button to the application's main page. On clicking this button, the background starts flashing between two colors at fixed intervals. This feature is implemented using Javascript, HTML, and CSS.

In the HTML file, a 'Start' button is added inside the 'background' div element. In the JS file, an event listener is added to the 'Start' button that initiates the flashing when clicked. The flashing functionality is implemented using setInterval to change the background color at a fixed interval.

The CSS file has been updated to style the added 'Start' button and to set the initial background color.

Issue: main.