feat: Update .gitignore and add GitHub Actions workflow
This commit adds the .idea directory to the .gitignore file to ignore
IntelliJ IDEA project files. It also introduces a GitHub Actions workflow
for publishing the project. The workflow will be triggered on each push
to the main branch. The files index.html, app.js, and style.css have also
been relocated to the src directory.